backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
langchain-community
langgraph
sqlalchemy
pydantic
python-dotenv
pytest
pytest-asyncio
faiss-cpu
numpy
